+++ weather@mockturtl/src/3_8/providers/met_norway/provider.ts
